* Jan Peter den Heijer <jpdenheijer@gmail.com>
The attached patch places the $root check before killing udev. The point is
this: It gives udev a chance to create the device. It also prints a message
saying "if you see the device being created here, try passing rootdelay=8 or
higher to the kernel"

also some minor edits:
- Use echo instead of message when dropping into the shell when $root isn't
  found. People will always want to see those message i guess...
- fixed a little typo : "use d echo * " is now "use echo * "

mkinitcpio-rev199-place-rootcheck-before-killing-udev.patch


git-svn-id: http://projects.archlinux.org/svn/initramfs/mkinitcpio@200 880c04e9-e011-0410-abf7-b926e227c9cd