rootdelay check: Fix compatibility with other sh implementations than dash and make it more reliable
The old implementation failed on bash and failed on both if rootdelay was not a number.
The logic is now as follows:
If "$rootdelay"="", then [ -z "${rootdelay}" ] is true
If not, then the part after the || is executed - if ${rootdelay} is not a number, or is <0, then [ ${rootdelay} -ge 0 ] fails,
and the test will return true due to the !. The previous implementation failed here, because [ ! ${rootdelay} -ge 0 ] returns false
on error, while ! [ ${rootdelay} -ge 0 ] returns true on error.