Create /dev/{null,zero,mem,console} devices when devtmpfs is missing
These devices used to be present because they were added in the image.
Since we now mount {dev,}tmpfs early in initramfs, the devices from
the image will be invisible. In the tmpfs-case, create the devices
manually so everything will be in order until udev is started, or
if udev isn't used at all.