Short circuit add module if it already exists in image
Rather than go through all the hoops, check if the module has already
been added by looking for its file path and skip the addition early if
we can. This helps especially in modules with many recursive
dependencies, and is most noticable when generating a full image with
many modules.

Signed-off-by: Dan McGee <dan@archlinux.org>