Don't print success/failure result on same line
Doing the `echo -n` call causes the shell to line-buffer the output
anyway, meaning one doesn't see the first part until 'SUCCESS' or
'FAILURE' is printed. This means the last message on the screen is
'Generating module dependencies' when in fact we are actually generating
the image, which can take a lot of time when XZ, etc. are used.

Print the status on two lines, ensuring buffering is not causing
problems with out output.

Signed-off-by: Dan McGee <dan@archlinux.org>