functions: cleanup and refactor add_binary
Modify the sed filter to remove PIC addresses and ignore the first line
which will always be the vdso. Also remove the tls filter, as glibc has
had threaded support mainlined for many years now. If a user has a glibc
this old, we don't have support for them elsewhere (udev, kernel).

We also cleanup the logical flow through this function and remove a lot
of cruft that would always return true, or that didn't quite make sense.

Signed-off-by: Dave Reisner <d@falconindy.com>