use bsdcpio to create images
This is a departure from using gen_init_cpio, which proved to be a huge
bottleneck in performance. Tests for existance change from being a call
to grep, to a simple shell test.

In the process, we have to modify the behavior of the -s option, and we
lose the -a option.

Signed-off-by: Dave Reisner <d@falconindy.com>