overhaul output, introducing color
Output style and coloring is borrowed from makepkg with minor
adjustments. Most instances where we encounter fatal errors are replaced
with a 'die' function to wrap this up neatly. In addition, we introduce
the -n flag, for no-color.

As a side effect, we need to source the functions file earlier so we
have the output functions available earlier.

Signed-off-by: Dave Reisner <d@falconindy.com>