add -t option to specify alternate build directory
This requires some refactoring of how we handle TMPDIR declaration and
creation, as well as protecting the way we do cleanup so that we don't
error out when cleanup is called prior to TMPDIR being declared.

We also add a shortcut to the build root based on the TMPDIR for
convenience.

Signed-off-by: Dave Reisner <d@falconindy.com>