functions: specify the delimiter to xargs
Without specifying this, xargs will split arguments on whitespace as
well as newlines, and will interpret quoting and backslashes. When the
delimiter is specified, every character is taken literally and only the
given delimiter in honored.

This sidesteps issues with broken modalias files as evidenced by a
MacBookAir3,1 or the bbs thread below:

https://bbs.archlinux.org/viewtopic.php?pid=971853

Also fixes FS#25450.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>