mkinitcpio: avoid using mountpoint
As it turns out, the mountpoint toy from util-linux isn't really at
feature parity with sysvinit's. In particular, and with respect to
v2.20:

- It throws false negatives when /proc isn't mounted. Rather, it should
  fall back on using stat(2), as is the basis for sysvinit's tool.
- when /proc is mounted with as source of 'none', it segfaults on a null
  dereference.

Patches have been sent upstream to resolve these issues, but in the
meantime, we're mounting /dev and /proc for very specific reasons.
Instead of relying on a fork/exec to figure out if these API FS's are
mounted, use a simple bash test to see if these specific needs are
available within the FS.

One of many possible fixes for FS#26344.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>