mkinitcpio: avoid stomping on TMPDIR from the environment
cc36db45 was a bit too strict about localizing variables and blew out
TMPDIR, which might be a legit variable sourced from the calling
environment. Restore this behavior with a bit of refactoring, and
additionally add in a check to make sure the temporary directory is
writeable before proceeding.

Fixes FS#26373.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>