move all binaries to /usr/bin
Add compat symlinks from /bin, /sbin and /usr/sbin, so nothing is lost.

This will make sure that it is not possibly to install two different
versions of the same binary in PATH.

The main usecase of this is that we want to be able to override any
symlink provided by busybox by adding a binary by the same name. With
this patch we don't have to worry in case the busybox symlink and the
binary we add are in different directories; the last binary to be added
always takes precedence.

Signed-off-by: Tom Gundersen <teg@jklm.no>