add_{file,symlink}: overwrite existing file
Now we fail silently if the file exists, which means there is no easy way to overwrite binaries.

I think it makes the most sense to do the following:
1) add all the busybox stuff as a basic system;
2) overwrite some of the links with any binaries we add to base hook;
3) allow users to write their own hooks to overwrite even more stuff.

This can now be done by this hook.

v2: make sure we are able to overwrite a symlink pointing to a directory

Signed-off-by: Tom Gundersen <teg@jklm.no>