mkinitcpio: include modules.{builtin,order} for depmod call
When present, modules.order will allow depmod to order the binary module
indicies for faster lookups and more deterministic behavior in resolving
aliases. We can discard this file from the buildroot after depmod is
called.

modules.builtin.bin is added as well (generated from modules.builtin),
to allow modprobe to not fail when a symbol is provided as a builtin.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>