mkinitcpio: remove --basedir option
This option is just a bad idea. Initramfs creation is too important to
get wrong, and running it from outside the root FS has too many gotchas,
the worst of them being:

- where do you pull hooks from?
- how do you resolve binary dependencies within the root?

In general, dealing with the extra luggage of the base directory makes the
codebase more complicated than it needs to be (see all the '_' prefixed
functions which are called from add functions). In favor of simplifying the
code, and making it more maintainable, kill this off and force the sane
option of chroot'ing into an install if the need arises.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>