lsinitcpio: extract the image to a tempdir
We read the image a number of times. Extract the image to a temp
directory so we can just extract it once and use the filesystem to our
advantage. This requires a temp dir which we can nuke on an EXIT trap,
but the whole operation is sped is a bit, especially for larger images.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>