init: add timestamp at start of initramfs if binary exists
This will allow us to tell how much time was spent in the initramfs,
which might be good for working on optimizations and making sure we
don't regress.

If a binary able to produce the timestamp is not installed this is a
noop.

I'll post separately an install hook to the systemd package that will
install systemd-timestamp. However, it would be trivial to replace this
by an independent implementation if people want to use this without
systemd.

With this patch I get the following on my (ancient) laptop:

  $ systemd-analyze
  Startup finished in 294ms (kernel) + 2684ms (initramfs) + 15741ms (userspace) = 18720ms

FWIW, 1 second is spent on inserting the needed modules, and 1.5 seconds
on mounting the rootfs (mounting btrfs is slow!).

Signed-off-by: Tom Gundersen <teg@jklm.no>
Signed-off-by: Dave Reisner <dreisner@archlinux.org>