Don't report success when builderrors > 0
Technically, we report success when we wrote an image to disk, but this
might not be obvious to users, even with a blaring warning right before
this line. If in the event builderrors is incremented at all, avoid
reporting any sort of "success", since there might be cause for concern.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>