merge most storage hooks into a single 'block' hook
As the autodetect hook gets better with every kernel release, we can
"afford" to remove the needless segregation of block device driver hooks
and simply merge all the "vanilla" or "unstacked" block device drivers.
This should make a lot of setups, like usb or virtio, easier as you
won't need to remember to add the hook.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>