functions: always return successfully if we ran a build hook
Removal of the old API (f85d28014134076) exposed the fact that a build
hook returning an error would increment the builderrors. This isn't
really wanted -- we don't want to declare a hard "pass/fail" for a hook,
as the functions we actually care about passing/failing are the add_*
functions which are of real concern (and already caught).

Signed-off-by: Dave Reisner <dreisner@archlinux.org>