mkinitcpio: avoid using 'file' to get kernel versions
The kernel defines an offset at a fixed location within the bzImage
which we can use to find the kernel version string. Since this is fairly
important, reimplement the algorithm (which file almost surely uses)
using low level tools to avoid the possibility that file may break or
change its output, critically wounding mkinitcpio.

Note that this change leaves 'file' in use for lsinitcpio compression
detection.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>