Add -l flag when using lz4 compression
This isn't documented, but it's needed since the kernel lz4 support is
based on the original streaming format and not the newer default. Use
-l to force the legacy codec.

Fixup lsinitcpio to support the older format, and, add a warning when
the newer format is detected. Who knows when this will be "resolved" ...

Upstream discussion: https://code.google.com/p/lz4/issues/detail?id=102