Handle softdeps in modules
modinfo outputs softdeps which we can add if they exist (failing
quietly on not-found). this obviates the need for some of our own
module quirks.

The output of modinfo can take many forms, as it's nearly freeform text
in the  .modinfo section of the kernel module. The basic form is:

  softdep:        pre: foo

But it might also be:

  softdep:        pre: foo bar post: baz

Or:

  softdep:        pre: foo
  softdep:        post: bar

So just parse the entire line, discarding anything that ends with a ':'.