Rework the install script a bit more
Added comments to make the flow more understandable. We now actively bail when
we encounter a kernel with no pkgbase. Also, if the preset does not exist yet,
and there is a .pacsave, we move that instead of copying from the template. Move
the preset check alongside the check if all was passed and finally, check if the
args array is not empty, before trying to run mkinitcpio.