Warn, if the preset file does not contain any presets
Previously, mkinitcpio would silently skip preset files that are
readable, but do not contain a PRESETS array at all or an empty one.

Printing a warning is chosen over printing an error / exit with failure
in order not to break (existing) scripts or use cases, in which multiple
preset files shall be processed, of which one might be empty
(intentionally?!).