functions: warn if the interpreter is missing when adding scripts with add_binary
Executables that are actually only scripts require the binary that is
specified in the shebang to be able run.

The `add_binary` function will only warn and not automatically include the
required interpreter.
This is because with things like `#!/bin/sh`, it may be better to use the
`base` hook that provides it instead of pulling in the `bash` binary.
Also it is possible that the interpreter is simply added after the binary
that requires it. That is why it says "Possibly missing".

Fixes #136.