functions: use cp instead of stat+install in add_file when mode is not specified
`cp` alone is considerably faster than running `stat` (to get the permissions)
and copying the file with `install`.
This basically replaces calls to two external commands with one.
The directory of the target file is created by `add_dir` when using `cp`.

When the mode is specified, `install` is used as before.