functions: always return 0 early when a non-binary file is passed to add_binary
138374f06b867b8c2f790bddb54b7ddd258ac493 added a condition to
`add_binary` that if the non-binary file does not have a shebang, the
function returns successfully before trying to parse sodeps. Before the
change, the function returned in all cases when the file was not a binary.

Bring back the previous behaviour of returning early on non-binaries.

Fixes 138374f06b867b8c2f790bddb54b7ddd258ac493