functions: do not create self-referential symlinks
Compare the symlink target with the destination in `add_file`.
If they match, simply place the file as a regular file in the specified
destination.

Additionally, move the "overwriting/adding file" message right before the
file is added. Showing the message before `add_dir` runs makes the verbose
output confusing, since it says that it adds the directory after the file
it is in.

Fixes https://bugs.archlinux.org/task/76812.
Fixes #154.