Put uki image into kernel-install staging folder
Systemd provides a kernel-install staging directory but it was not used
because the artifacts were not automatically copied from the staging
location to the final EFI folder. Systemd 253 includes a new
[90-uki-copy.install](https://github.com/systemd/systemd/blob/main/src/kernel-install/90-uki-copy.install)
script which expects the image in the staging directory and copies it
to EFI/Linux. The uki-copy script will fail if the file does
not exist.

To fix this the uki image is now written to the staging folder and the
uki-copy script puts it afterwards in the right location. This will only
work with systemd 253 and later.

Systemd commit: https://github.com/systemd/systemd/commit/0ccfd3564b2532a4da6526a9e030362c4a142b77
Systemd bug: https://github.com/systemd/systemd/issues/26477

Fixes #163