install/keymap: fix do not pass empty string to loadkeys
If $uc is not set, an empty string gets passed as an argument to loadkeys
 and this is not what it expects.

Place an arguments that will always be there in an array and append to
it if necessary.

This could be considered a minor issue since it only happens on non-UTF-8
locales and there is no good reason to use anything like that.

Fixes 1add8bb75301319a117c8ef98a9b1df67162074d
Fixes https://bugs.archlinux.org/task/77949