uki: add IMAGE_ID, IMAGE_VERSION and VERSION_ID to .osrel
This adds IMAGE_ID and IMAGE_VERSION to the .osrel section when built
from a preset and neither value exists in os-release. VERSION_ID is
always added as the kernel version if not in os-release.

The systemd-boot menu looks like this with this patch:
Arch Linux (linux)
Arch Linux (linux-fallback)
Arch Linux (linux-lts)
Arch Linux (linux-lts-fallback)