install/autodetect: retry looking at uevents
If the first time does not succeed, retry again, but only once. This will
hopefully avoid a race condition with udev.

This is a best-effort workaround for what is assumed to be a bug in
`udevadm settle` where the command exits while the devices have not yet
settled. See https://gitlab.archlinux.org/archlinux/mkinitcpio/mkinitcpio/-/issues/177
and https://bugs.archlinux.org/task/77789 for details.