install/autodetect: match on both MODALIAS and DRIVER
There are a handful of drivers which may be loaded but are missing a
MODALIAS entry, so match on both.

For example: on a couple of unrelated instances people have reported
that mkinitcpio is not adding the nvme driver to the initrd. As per
"modinfo nvme" the driver fits this case perfectly.

On my machine, this gets us from 54 unique modules to 62.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>