install/systemd: stop installing systemd-vconsole-setup.service
After [1] was merged, systemd-vconsole-setup.service is included to the
initramfs unconditionally. If the sd-vconsole hook is not included, it
leads to an error on boot, because the loadkeys and setfont binaries are
missing:

    systemd[1]: systemd-vconsole-setup.service: Main process exited, code=exited, status=1/FAILURE
    systemd[1]: systemd-vconsole-setup.service: Failed with result 'exit-code'.
    systemd[1]: Failed to start Virtual Console Setup.
    systemd-vconsole-setup[131]: /usr/bin/loadkeys failed with exit status 1.

Fix it by reverting the part of [1] that added
systemd-vconsole-setup.service, which now should be installed by the
sd-vconsole mkinitcpio hook.

[1]: https://gitlab.archlinux.org/archlinux/packaging/packages/systemd/-/commit/4ae88c260a0aeec89956063ccfb887a416f40f52

Fixes: b0beb7b4451d89676b50048a74c6c5690b2fcaf1 ("upgpkg: 254rc3-1: new upstream pre-release - testing only!")
Closes: https://gitlab.archlinux.org/archlinux/mkinitcpio/mkinitcpio/-/issues/247