50-mkinitcpio.install: pass path to kernel image to mkinitcpio
When kernel-install is used with a kernel image in a non-standard
location, mkinitcpio can not find it.
Instead pass the full path to the image as the kernel paramater so
mkinitcpio can easily read the image from there.

(cherry picked from commit abaebdaa4de2e6dd2689f9405411cc7e430f4766)