test: use system library loading mechanism instead of absolute paths
This resolves system libraries automatically from the canonical paths,
or falls back to `BATS_LIB_PATH`.

Previously this was not possible as bats plugins were not packaged at
the appropriate path in /usr/lib/bats/* which resulted in the plugin
auto loading not being able to resolve the libraries. This has been
fixed now, so lets use the recommended way to load libraries.

https://bats-core.readthedocs.io/en/stable/writing-tests.html#bats-load-library-load-system-wide-libraries

Signed-off-by: Levente Polyak <anthraxx@archlinux.org>