functions: support staging an early image in the buildroot
This splits the buildroot into two dirs: early and root. early is used
for things that should be loaded before the initramfs, like microcode.
root is the same as it was before.

because of the logic changes in the function, initialize_buildroot no
longer unconditionally creates a working directory in /tmp.

test cases for initialize_buildroot are also added.