functions: add_file: if the parent directory is a symlink, add the directory symlink
When a file that exists in a symlinked directory needs to be added,
copy the full target directory and add the directory symlink.
This is only done for the immediate parent directory of the file. For
any symlinks higher up in the file path, we simply hope they never
become an issue.

This avoids bloating the image by adding including duplicates of a file.

Fixes https://gitlab.archlinux.org/archlinux/mkinitcpio/mkinitcpio/-/issues/238